requirements.txt
flask==3.1.0
line-bot-sdk==3.11.0
anthropic==0.49.0
python-escpos==3.1
python-dotenv==1.1.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
